app/requirements.txt
requests
websocket-client
Flask
python-dateutil